requirements.txt
appdirs==1.4.3
bcrypt==3.1.4
boto3==1.9.38
django-environ==0.4.5
dj-database-url==0.5.0
dj-static==0.0.6
django-storages==1.7.1
Django==2.1.7
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
djangorestframework==3.9.1
gunicorn==19.9.0
newrelic==4.8.0.110
olefile==0.46
packaging==18.0
Pillow==5.3.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
python-dotenv==0.10.3
pyparsing==2.3.1
pytz==2018.9
six==1.12.0
